add_forge_now: Add last_{moderator, modified_date} fields to Request
Previously, the last moderator and modification date for an Add Forge Now
request were retrieved by the model serializer.

While it is enough for display, it prevents to sort the requests by last
modification date when using datatables.

So prefer to add the last_moderator and last_modified_date to the Request
model instead and update their values when a new RequestHistory model
is created.

When running migrations, existing requests stored in production database
will have the values of these new fields set according to their history.