save_origin_webhooks: Add push webhook receivers for popular forges
Add new Web API endpoints accepting only POST requests coming from push
webhooks of the following popular forges or their instances:

- Bitbucket
- Gitea
- GitHub
- GitLab
- SourceForge

There is one API endpoint per forge type enabling to request or update the
archival of a repository when new commits are pushed to it through the Save
Code Now service. Each endpoint simply processes the webhook JSON payload
sent by a forge in order to extract the repository URL and the visit type
in order to create a new save request for the repository.

Related to T4548