save_code_now: Allow request creation if previous for origin is running
Previously, the save code now service was not allowing to create a new
request for an origin if a previously created one for the same origin
was currently running.

However, new commits might have been pushed to the origin while a loading
task is running so this restriction does not make sense, especially for
the new forge webhook receivers feature.

Related to T4548