metadata-search: Return swh-search even when missing from idx_storage.origin_intrinsic_metadata
This is needed, because swh-search may now return results based on extrinsic metadata,
in addition to intrinsic metadata.

I do not want to query idx_storage.origin_extrinsic_metadata here, because it
is not clear how to merge with the existing data structure.

Additionally, I do not think anyone relies on the metadata returned by this
endpoint because it is undocumented and rather inflexible. Instead, I would
like to deprecate returning metadata from this endpoint altogether, as there
is a more appropriate endpoint to get metadata once you have the origin URL.