browse: Display branches names in snapshot releases view
Previously the snapshot releases view was only displaying the releases
names and not the branches names targeting those.

But it exist cases where multiple branches target the same release in
a snapshot so also display branches names in the releases table in order
to avoid confusion about duplicated releases.

Related to T4686