webapp/readme-rendering: Fix markdown relative link to archived object
Some README files written in markdown can contain relative links to files
or directories reachable from the directory containing the README.

The javascript library we are using to render markdown to HTML, showdown,
transforms such links to invalid browse URLs leading to 404 errors.
Those URLs are in the form /browse.*/directory/<path>/ while the correct
ones should be in the form /browse.*/directory/?path=<path>.*

So implement a post process step after rendering markdown to HTML in
order to fix this type of links and browse the targeted contents or
directories.

Related to #4331