utils: Remove caching of get_deposit_lists and use pagination
The returned list of deposits might have a large size and could exceed
the memcache size limit. That function is only used by the coverage widget
to count deposits and these counts are already cached so no need to cache
the deposit lists.

Also requests deposits data in a paginated way.