save_code_now: Add endpoint to receive origin visit webhook messages
In order to update the statuses of save code now requests in a push manner
instead of the current pull one, add a new HTTP endpoint to receive webhook
messages about origins being visited by the SWH archive.

When the visit of an origin results from a Save Code Now request, the status
of the request is then updated when receiving webhook messages notifying
the new status of the visit.

Related to swh/meta#4836.