cypress: Replace cypress-cloud by cypress-split for parallel tests
Cypress >= 13.3 blocked the use of sorry-cypress and cypress-cloud in the
cypress binary so we can no longer use these tools to execute parallel
cypress tests on Jenkins.

Fortunately, it exists another tool called cypress-split that can be used
for the same purpose without relying on a tests orchestration server.

So replace the use of cypress-cloud by cypress-split to still benefit of
parallel cypress tests in CI jobs.