browse/assets/swhid-utils: Fix qualified SWHID URLs set by addLinesInfo
When checking SWHID context checkboxes or selecting lines in a content, the
addLinesInfo function updates the URLs to browse qualified SWHIDs but it was
not setting the percent escaped URLs required to properly browse SWHIDS having
special characters in their qualifiers.

So ensure to set proper browse URLs to not break the browsing of such SWHIDs.