save_code_now: Rename "not yet scheduled" loading status to "pending"
The statuses of save code now requests will be updated in a push way
through use of webhooks soon in production and as a consequence the
"scheduled" loading task status will no longer be reported as webhooks
first reported status is "running".

So rename the "not yet scheduled" status to "pending" to allow better
transitions for loading task statuses.

Previous status transitions were:  not created -> not yet scheduled
-> scheduled -> running -> (succeeded|failed).

New status transitions are: not created -> pending -> running
-> (succeeded|failed).

Related to swh/infra/sysadm-environment#5165.