Stop reporting RemoteException to Sentry
In practice, RemoteException happens only when a backend service has an internal
error (which is already logged in its own Sentry project) or a query timeout;
and TransientRemoteException only on temporary 502 errors.

Neither are helpful, and they clutter Sentry by bursting a dozen such errors
a couple of times a week.