auth: Fix logout with Django >= 5 default authentication backend
Since Django 5.0, the logout view of the default authentication backend
no longer accepts GET requests but only POST ones so adapt swh-web
to take this change into account and enable proper logout operation
when using the default authentication backend with Django >= 5 (it is
used when running cypress tests or executing the webapp in docker for
instance).

Also fix a couple of issues related to login/logout in templates
and improve styling a bit.