browse/snapshot_context: Do not store branches messages in cache
To avoid retrieving the list of branches in each page when browsing a
snapshot, it is put in cache.

However the messages associated to branches are also stored in cache
but those can have a large size and the cache size limit might be
reached.

As the messages of the branches are only exploited in the branches
and releases view, ensure to not store them in get_snapshot_content
cache to avoid MemcacheServerError being thrown due to a too large
value to cache.