browse: Improve browsing an origin with multiple visit types
Some origins in the archive can have multiple visit types since
recently and could lead to confusing display in browse interface
as the latest visit is selected by default.

So add a select element at the top right part of the UI enabling
to switch the visit type for the browsed origin and update links
by adding the visit_type query parameter to keep visit type context
when navigating.

Related to #4786.