Add an error page to handle masked objects
We catch MaskedObjectException at the webapp level and display a
specific error page. The page lists the opaque identifier of the masking
requests and the legal email address where to send inquiries.

We also add support for MaskedObjectException in API responses.

Thanks to @anlambert for deduplicating the error templates and fixtures.
Thanks to @vlorentz for improving the wording of the error page.