add_forge_now: do not set a default forge type
Uses a hidden disabled selected option with an empty value to do this.

Prevents accidentally submitting forges as the first forge type,
which is currently bitbucket, which is rarely used on the web.

The AFN submission API already rejects an empty forge type,
giving a reasonably good error to the end-user.