auth: Add new migration to create webapp permissions in database
Add a new migration file for the swh.web.auth django application enabling to
persist webapp special permissions to database in order to be able to use them
when the webapp relies of the default django authentication backend for login
operations (mirror version of swh-web for instance that does not use Keycloak
for authentication).

The permissions can then be set for a particular user using the django admin
views by a django user having the adequate permissions to modify user accounts.

Some tests were modified as the reset_sequences parameter of django_db pytest
mark cannot be used as it breaks persisted permissions.