utils: Update test checking characters escaping in a SWHID URL
swh-model was updated to fix deserializing of origin qualifier in
a QualifiedSWHID instance which breaks a test of the webapp.

Align test with changes in swh-model and improve it to check that
a qualified SWHID containing special characters that must be escaped
can be successfully resolved to a browse URL but also that the latter
can be successfully browsed.