save_bulk: Add Web API endpoint to get feedback about a sent request
Add Web API endpoint /origin/save/bulk/request/(request_id)/ to get
feedback about a previously sent save bulk request.

The endpoint requires authentication and special permission to be
queried, also only the user that submitted the request or SWH staff
members can query it.

It returns for each submitted origin the following info:

- the origin URL
- the visit type
- the origin status regarding its processing by the bulk save lister,
  either pending, accepted or rejected
- the last scheduling date if any
- the last visit date if any
- the last visit status if any
- the last snapshot SWHID if any
- the rejection reason if the origin was rejected by the lister
- the browse URL if the origin was successfully visited

These info are returned in a paginated way as the number of submitted
origins can be large.

Related to #4802.