webapp/assets: Add javascript function adding anchors to headings
Add javascript function based on the use of the anchorjs package enabling
to add anchors to headings. It enables users to easily get links to parts
of a page by hovering mouse cursor on page headings.

Use it in homepage, api pages but also in pages rendering a README file
from the archive.