Add a sync_masking_mailmaps admin command
This command will copy all the active mailmaps in the masking proxy DB
which dsn is given as argument.

It is similar to the (now deprecated) sync_mailmaps command, but since
it does not modify the Person table of the storage any more, it does not
need the timestamp-based detection to determine which entries are to be
enabled or disabled; it simply rewrite the whole displayname table
every time it's called, if need be.

A '--force' option allows to enforce rewrite the whole table even if no
update has been detected in the displayname database.