webapp/assets: Use showdown-highlight from npm instead of vendoring it
showdown-highlight is a plugin for showdownjs enabling to highlight a code
block with highlightjs when rendering markdown to html.

There was some issues at the time to use the npm version but it is no longer
the case so fetch that code from the npm registry instead of vendoring it
in swh-web frontend code.

That updated version also fixes some rendering issues related to code blocks
with no explicitly declared programming language.