browse: Add new side tab to hold the Citations Web UI
In order to get an easy access to the incoming software citations
Web UI for the currently browsed software objects, add a new side
tab whose handle is located below the Permalinks tab one.

When clicking on any tab handles, both tabs are slided to the left
and the current displayed one can be switched by clicking on its
handle once the tabs opened.

Also update guided tour to focus on the citations tab after the
permalinks one, it should be completed when the citations Web UI
will be integrated.

Related to product-management/swh-archive-website#4.