browse/sidetabs: Integrate software citations Web UI
Integrate Web UI to generate software citations in BibTex format
in the new Citations side tab on the right.

The Web UI is similar to the one for Permalinks with the difference
that BibTex citations are generated on the fly through Web API calls
when selecting an object type.

This new Citations Web UI and its sidetab are not integrated by
default and can be activated through configuration by setting the
activate_citations_ui entry to True.

Related to product-management/swh-archive-website#4.