Adjust wording of the initial AFN admin contact
Drop the mention of the request since it can cause admin concerns.

Switch from the "forge" terminology to "code" and site" since
"forge" is outdated terminology since the advent of GitHub,
which did not keep up the tradition of old code forges.

Mention that the archiving will be regular too.