citation: Ensure source SWHID is available when generation failed
When a parsing error occurs when attempting to generate a citation,
store error message in Citation dictionary and return an error response
enriched with SWHID of citation metadata source file.

It enables to get feedback to users about errors in the citations Web UI
while offering link to the bogus metadata file.