browse: Fix redirect issue in search interface
When searching for an origin URL, the search interface redirects
to its browse interface if the origin URL exists in SWH database.

Nevertheless, it the origin could not be archived, the redirection
was still done leading to a 404 error so ensure to check the origin
has a snapshot before redirecting.

Fixes #4828.