utils/archive: Fix origin_visit_find_by_date search behavior
When the greater_or_equal parameter is set to True and there is
no visit date greater or equal to the one passed as parameter,
this function should return None.

This fixes generated browse links in the Save Code Now Web UI
when the same origin was submitted multiple times for archival.