save_code_now: Fix status of request when celery worker was killed
It can happen that a celery worker executing a save request gets killed
for various reasons: OOM, no more disk space, ...

Previously the save request status was computed from the origin visit
status but if the worker got killed in flight, the final visit status
was not set and thus the save request was erroneously marked as still
running while it has failed.

So ensure to detect the killed worker situation by also checking task
statuses in scheduler database when task is detected as running from
the latest origin visit status.

Related to swh/devel/swh-web#4835.