archive_coverage: Fix and improve performance of deposit counts
Previous implementation was quite naive as it was retrieving all
deposits data to count them but the number of deposits grew a lot
recently since zenodo using the service. Also the counters were
not accurate as metadata deposits and failed ones were considered.

Prefer to count them using swh-search instead which is performant
and accurate.