save_bulk/api_views: Bump per_page value for the save bulk lister task
When creating the save bulk lister task after all submitted origins were
processed, do not set the per_page value to 10 and use the default value
of the lister (1000).

This should speedup the listing a bit as less HTTP requests will be sent
to the webapp to fetch the list of origins to process.