bootstrap: Fix modal behavior when animations are disabled
When animations are disabled through the desktop environment settings,
bootstrap CSS animations are also disabled through the CSS media feature
prefers-reduced-motion that is set to reduce. In that case special CSS
rules are used to disable animations.

Nevertheless, bootstrap modals having the CSS class fade see their display
attribute set to block when animations are disabled which turns every modals
present in the DOM to be visible, covering every other DOM elements and thus
preventing to interact with them which makes the displayed page unusable.

Ensure every modals are not displayed by default and add event handlers when
modals are displayed or hidden to restore expected behavior.

Fixes #4841.