browse: Add optional HTML or Markdown preview in content views
When browsing an HTML or Markdown document, users now have the
possibility to request a rendering preview of such document by
clicking on a dedicated radio button.

For obvious safety concerns, the HTML documents are sanitized
in the first place to prevent any XSS exploits and are rendered
in a dedicated iframe.

Nevertheless, the loading of CSS files and image assets are
allowed and their bytes will be fetched from the archive if
available. To allow such feature, DOMPurify hooks have been
updated and an internal endpoint has been modified to use url
arguments instead of query parameters.

Navigation between relative documents in the archive is also
functional.

To directly render the preview when loading the page, the query
parameter named show_preview must be present in URL and set to
true.

Fixes #4840.