browse/content: Allow math display when rendering markdown to html
When a user requests the rendering of a markdown file to HTML,
embedded math formula (using for instance LaTeX syntax) will
be nicely rendered thanks to the use of the MathJax library.

Move function to load and init MathJax in its own javascript
chunk in order to use it inside the iframe rendering HTML
converted from markdown.