browse/origin-visits: Fetch up to 10000 visits to avoid timeouts
Some origins can have a really large amount of visits, up to the million,
which could cause timeouts as the webapp attempted to fetch all visits
regardless their amount.

From now on, fetch up to 10000 recent visits for an origin and display a
warning message when a partial list of visits was fetched.

Related to #4847.